+++ libraries/ADC081S/CHANGELOG.md
@@ -6,12 +6,18 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
-## [0.1.1] - 2024-01-10
+## [0.1.2] - 2024-01-12
+- add support for **ADC101S** 10 bit series
+- add support for **ADC121S** 12 bit series
+- update readme.md
+- minor edits
+
+
+## [0.1.1] - 2024-01-12
 - add **void lowPower()** function
 - add **void wakeUp()** function
 - add **bool isLowPower()**
 - add example
-
 
 ## [0.1.0] - 2024-01-10
 - initial version.